item endpoint calls getAll
