+++ src/app/api/equipment/route.ts
@@ -19,7 +19,7 @@ export async function GET(request: Request) {
   const potionOfExpeditiousRetreat: Equipment = {
     name: 'Potion of Expeditious Retreat',
     description:
-      'The stopper for a potion of expeditious retreat is crafted to easily snap open in dire circumstances. When you drink this potion, you become @condition:fleeing for 1 minute, and you gain a +40-foot status bonus to all your Speeds for the duration as long as you are fleeing. You immediately Stride.',
+      'The stopper for a potion of expeditious retreat is crafted to easily snap open in dire circumstances. When you drink this potion, you become @condition:fleeing@ for 1 minute, and you gain a +40-foot status bonus to all your Speeds for the duration as long as you are fleeing. You immediately Stride.',
     price: [{ value: 3, type: 'gp' }],
     bulk: 'L',
     level: 1,
